requirement.txt
anyascii==0.3.1
attrs==21.4.0
beautifulsoup4==4.11.1
blinker==1.6.2
blis==0.7.9
boto3==1.26.133
botocore==1.29.133
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
contractions==0.1.73
cryptography==40.0.2
cymem==2.0.7
docx2txt==0.8
filelock==3.12.0
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.2
murmurhash==1.0.9
mysql-connector-python==8.0.33
nltk==3.8.1
numpy==1.24.3
packaging==23.1
pafy==0.5.5
pandas==2.0.1
passlib==1.7.4
pathy==0.10.1
pdfminer.six==20221105
pdfminer3==2018.12.3.0
Pillow==9.5.0
plotly==5.14.1
pocketsphinx==5.0.0
preshed==3.0.8
protobuf==3.20.3
pyahocorasick==2.0.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
PyMySQL==1.0.3
pyresparser==1.0.6
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
regex==2023.3.23
requests==2.29.0
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
snscrape==0.6.2.20230320
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.10.0
srsly==2.4.6
tenacity==8.2.2
textsearch==0.0.24
thinc==8.1.9
threadpoolctl==3.1.0
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wasabi==1.1.1
Werkzeug==2.3.3
